tweet_collector/requirements.txt
requests
pymongo
mastodon.py
bs4